+++ SlashBar/Modules/Native/HotkeyNative.cs
@@ -2,12 +2,9 @@ namespace SlashBar.Modules.Native;
 
 public static class HotkeyNative {
 
+    public const uint ModAlt = 0x0001;
     public const uint ModControl = 0x0002;
     public const uint ModShift = 0x0004;
-
-    public const uint VkSpace = 0x20;
-    public const uint VkQ = 0x51;
-    public const uint VkA = 0x41;
 
     public const int WmHotkey = 0x0312;